simplify code a bit
