+++ lib/datadog/ci/auto_instrument.rb
@@ -1,5 +1,4 @@
 require "datadog/ci"
 require "datadog/ci/contrib/contrib"
-require "datadog/ci/contrib/kernel"
 
 Datadog::CI::Contrib.auto_instrument!
