+++ lib/datadog/ci/contrib/cucumber/integration.rb
@@ -28,10 +28,6 @@ module Datadog
             super && version >= MINIMUM_VERSION
           end
 
-          def requires
-            ["cucumber"]
-          end
-
           def auto_instrument?
             true
           end
